% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mesonet_data.R
\name{getMesonetData}
\alias{getMesonetData}
\title{Pull data from meterologic stations in the mesonet network}
\usage{
getMesonetData(token, mesoId, timeBgn, timeEnd)
}
\arguments{
\item{token}{Personalized Mesonetwork token.  this is needed to download data}

\item{mesoId}{Mesonet ID for site of interest}

\item{timeBgn}{A YYYY-MM-DD hh:mm:ss formated start time, in UTC.}

\item{timeEnd}{A YYYY-MM-DD hh:mm:ss formated end time, in UTC.}

\item{timeScale}{The time scale of measurements to return.
Enter 'monthly', 'daily', 'hourly', or 'subhourly'.}
}
\value{
Data frame of mesonet data
}
\description{
Provide a valid mesonet ID for a site in the mesonetwork; your personal
mesonetwork data token, and requested begin and end times in POSIX format,
and a data frame will be returned to the global environment.
}
\examples{
\dontrun{
token="16088448e1b149509e45e401196106f0"
mesoId="D4130"
timeBgn="2019-01-01 00:00:00"
timeEnd="2019-02-01 00:00:00"

getMesonetData(timeScale = timeScale, TimeBgn = TimeBgn, TimeEnd = TimeEnd, mesoID=mesoID, token=token)
}

}
\author{
Josh Roberti \email{jroberti@battelleecology.org}\cr
}
\keyword{climatology,}
\keyword{data,}
\keyword{ecology,}
\keyword{mesonet,}
\keyword{metScanR}
\keyword{meteorology,}
\keyword{weather,}
