% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_single_acis.R
\name{getSingleACIS}
\alias{getSingleACIS}
\title{Download Data Served through ACIS?}
\usage{
getSingleACIS(sid, start_date, end_date,
  vars = "mint,maxt,avgt,pcpn,snow,snwd,obst")
}
\arguments{
\item{sid}{Parameter of class character. The (GHCND) site ID of the site data should be downloaded for.}

\item{start_date}{Parameter of class character. The year-month-day (e.g. "2017-01-02") of the last
date to get data for.}

\item{end_date}{Parameter of class character. The year-month-day (e.g. "2017-01-02") of the last
date to get data for.}

\item{vars}{Optional. A string of element codes, separated by commas only. Defaults to "mint,maxt,avgt,pcpn,snow,snwd,obst"}
}
\value{
Writes results file to the specified directory.
}
\description{
GETUSOMEDATA
}
\examples{
Currently none
}
\seealso{
Currently none
}
\author{
Cody Flagg \email{cflagg@battelleecology.org}\cr
}
\keyword{commissioning}
\keyword{data}
\keyword{gaps,}
\keyword{process}
\keyword{quality,}
